fix: [All TESTS PASSED] Fixed issue where calling Dispose of multiple encrypt/decrypt Streams.
Changed the implementation of method "DisposeCreatedStreams". Call Dispose only on the first stream in the stream chain.
Changed "leaveOpen" values on SymmetricAlgorithmAdapter.SystemCryptography.AES to false.

NOTE:
The crux of the problem is that the streams in the chain have Dispose called multiple times. However, it is not guaranteed (I think) that when a stream is called Dispose, it will recursively call Dispose on the streams related to itself.
There are prerequisites for this change. It is assumed that by calling Dispose for only the first stream, Dispose for all streams will be called recursively.

Therefore, method "DisposeCreatedStreams" is an unstable implementation. The problem may become apparent when the number of concrete classes of interface ISymmetricAlgorithmAdapter used by class SymmetricEncrypterBase increases.