test: Adjusted time limit of test "ExtractWithLongChain_CompleteWithinTime".
Test "StashWithLongChain_CompleteWithinTime" has 5 seconds. X is executed only once at the time of creation.
Test "ExtractWithLongChain_CompleteWithinTime" has 1 second. Since this is executed every time, the processing time must be short.

As a result of clearly defining the measurement period, it was determined that improvement of performance was unnecessary at this time.