refactor: Refactored the constructor of class "BasicEdDataLogicFactory".
Now use the value of a class property when instantiating a logic object.
The default values for these properties are defined as in-class constants.
Values defined as constants are adjusted for processing load and security considerations.
Alternatively, the value provided in the constructor argument is used instead.

This change was made to reduce the processing time of the test process.
Therefore, there is no need to change the client code.