perf: The load was significantly reduced by logic setting.
The processing was a problem with practicality, so the load was significantly reduced. If there are more indexes to decode, this adjustment will still require lowering the processing load.
The impact of the database enlargement was beyond expectations. Therefore, the FakeInsertion value has also been reduced.

In any case, I feel that there is a need for an interface to dynamically set the load. It is impossible to balance the load that can be adapted to any environment while improving security.